+++ lib/install/dotfiles/wallpaper.sh
@@ -18,11 +18,11 @@ else
         _writeMessage "You can download and install additional wallpapers from https://github.com/mylinuxforwork/wallpaper/"
         echo
         if gum confirm "Do you want to download the repository?" ;then
-            if [ -d ~/Downloads/wallpaper ] ;then
-                rm -rf ~/Downloads/wallpaper
+            if [ -d $download_folder/wallpaper ] ;then
+                rm -rf $download_folder/wallpaper
             fi
-            git clone --depth 1 https://github.com/mylinuxforwork/wallpaper.git ~/Downloads/wallpaper
-            rsync -a -I --exclude-from=$includes_directory/excludes.txt ~/Downloads/wallpaper/. ~/wallpaper/ &>> $(_getLogFile)
+            git clone --depth 1 https://github.com/mylinuxforwork/wallpaper.git $download_folder/wallpaper
+            rsync -a -I --exclude-from=$includes_directory/excludes.txt $download_folder/wallpaper/. ~/wallpaper/ &>> $(_getLogFile)
             _writeLogHeader 1 "Wallpapers from the repository installed successfully."
         elif [ $? -eq 130 ]; then
             exit 130