+++ lib/install/packages/fonts.sh
@@ -6,13 +6,13 @@ if [ -d ~/.local/share/fonts/JetBrainsMonoNerd ]; then
 else
     # Download JetBrainsMonoNerdFonts
     DOWNLOAD_URL="https://github.com/ryanoasis/nerd-fonts/releases/latest/download/JetBrainsMono.tar.xz"
-    curl -o $HOME/Downloads/JetBrainsMono.tar.xz -OL "$DOWNLOAD_URL" &>> $(_getLogFile)
+    curl -o $download_folder/JetBrainsMono.tar.xz -OL "$DOWNLOAD_URL" &>> $(_getLogFile)
 
     # Install the fonts
     mkdir -p ~/.local/share/fonts/JetBrainsMonoNerd &>> $(_getLogFile)
 
     # Extract the new files into the JetBrainsMono folder and log the output
-    tar -xJkf $HOME/Downloads/JetBrainsMono.tar.xz -C ~/.local/share/fonts/JetBrainsMonoNerd
+    tar -xJkf $download_folder/JetBrainsMono.tar.xz -C ~/.local/share/fonts/JetBrainsMonoNerd
     _writeLogTerminal 0 "JetBrainsMono fonts installed"
 fi