+++ setup-arch.sh
@@ -160,23 +160,6 @@ fi
 sudo pacman -Sy
 echo
 
-# pacman
-if [ -f /etc/pacman.conf ] && [ ! -f /etc/pacman.conf.ml4w.bkp ]; then
-    echo -e "\033[0;32m[PACMAN]\033[0m adding extra spice to pacman..."
-
-    sudo cp /etc/pacman.conf /etc/pacman.conf.ml4w.bkp
-    sudo sed -i "/^#Color/c\Color\nILoveCandy
-    /^#VerbosePkgLists/c\VerbosePkgLists
-    /^#ParallelDownloads/c\ParallelDownloads = 5" /etc/pacman.conf
-    sudo sed -i '/^#\[multilib\]/,+1 s/^#//' /etc/pacman.conf
-
-    sudo pacman -Syyu
-    sudo pacman -Fy
-
-else
-    echo -e "\033[0;33m[SKIP]\033[0m pacman is already configured..."
-fi
-
 # Install required packages
 echo ":: Checking that required packages are installed..."
 _installPackages "${packages[@]}"