+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 2.9.8
 - SwayNC added as new notification app and center with additional features.
 - kitty, nvim, .bashrc and .zshrc requires user confirmation for overwriting.
 - All required additional downloads required for the installation will be stored in .ml4w folder
+- Walcord support in wallpaper.sh script
 
 Version 2.9.7.4
 --------------------------------------------------------